refactor(harness-skills/store): extract tests via #[path]
Phase 0 PR-8 of the architecture decoupling roadmap. Third clean U-16
victory: extraction brings store.rs fully under the 800-line ceiling.

Net effect:
  harness-skills/src/store.rs:        1369 -> 757 lines (-612)  ← under U-16 ✓
  harness-skills/src/store_tests.rs:   NEW   612 lines

Same pattern: `#[cfg(test)] #[path = "store_tests.rs"] mod tests;`.

Verification:
- cargo test -p harness-skills --lib:
    54/54 passed (covers all skills::store::tests::* including
    record_use_persists_sidecar, persisted_skill_shadowing_builtin_*,
    discover, watcher, validation cases)
- RUSTFLAGS=-Dwarnings cargo check --workspace --all-targets: clean
- cargo fmt --all -- --check: clean
